Add verification of task ownership

+++ src/main/java/com/petize/todolist/domain/services/TaskService.java
@@ -13,6 +13,8 @@ import com.petize.todolist.domain.models.User;
 import com.petize.todolist.domain.models.enums.Priority;
 import com.petize.todolist.domain.models.enums.TaskStatus;
 import com.petize.todolist.domain.repositories.TaskRepository;
+import com.petize.todolist.domain.services.auth.AuthenticationService;
+import com.petize.todolist.exceptions.AccessDeniedException;
 import com.petize.todolist.exceptions.EntityNotFoundException;
 import com.petize.todolist.exceptions.NotCompletedSubtaskException;
 
@@ -23,6 +25,7 @@ import lombok.AllArgsConstructor;
 @AllArgsConstructor
 public class TaskService {
     
+    private final AuthenticationService authenticationService;
     private final TaskRepository taskRepository;
     
     public Task save(@Valid Task task) {
@@ -37,16 +40,21 @@ public class TaskService {
     }
 
     public Task getById(Integer id) {
-        return taskRepository.findById(id).orElseThrow(
+        var task = taskRepository.findById(id).orElseThrow(
             () -> new EntityNotFoundException(Task.class, id)
         );
+
+        validateTaskOwnership(task);
+        return task;
     }
 
-    public Page<Task> getAllWithFilters(User user, LocalDate dueDate, TaskStatus status, Priority priority, Pageable pageable) {
+    public Page<Task> getAllWithFilters(LocalDate dueDate, TaskStatus status, Priority priority, Pageable pageable) {
+        var user = authenticationService.getUserAuthenticated().getUser();
         return taskRepository.findAllWithFilters(user, dueDate, status, priority, pageable);
     }
 
-    public void updateStatus(Task task, TaskStatus status) {
+    public void updateStatus(Integer id, TaskStatus status) {
+        var task = getById(id);
         
         if (status == COMPLETED && containsNotCompletedSubtasks(task)) {
             throw new NotCompletedSubtaskException(task.getId());
@@ -63,6 +71,14 @@ public class TaskService {
     }
 
     public void delete(Task task) {
+        validateTaskOwnership(task);
         taskRepository.delete(task);
+    }
+
+    public void validateTaskOwnership(Task task) {
+        var user = authenticationService.getUserAuthenticated().getUser();
+        if(!task.getUser().equals(user)){
+            throw new AccessDeniedException("This user is not allowed to access this task");
+        }
     }
 }
